Task-2/requirements.txt
py-algorand-sdk>=2.0.0
python-dotenv>=0.19.0
